Rename listcontainer to tripcontainer
